+++ apps/web/pages/api/book/event.ts
@@ -1,0 +1,73 @@
+import { checkBotId } from "botid/server";
+import type { NextApiRequest } from "next";
+
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
+import handleNewBooking from "@calcom/features/bookings/lib/handleNewBooking";
+import { checkRateLimitAndThrowError } from "@calcom/lib/checkRateLimitAndThrowError";
+import getIP from "@calcom/lib/getIP";
+import { piiHasher } from "@calcom/lib/server/PiiHasher";
+import { checkCfTurnstileToken } from "@calcom/lib/server/checkCfTurnstileToken";
+import { defaultResponder } from "@calcom/lib/server/defaultResponder";
+import { CreationSource } from "@calcom/prisma/enums";
+
+async function handler(req: NextApiRequest & { userId?: number }) {
+  const userIp = getIP(req);
+
+  if (process.env.NEXT_PUBLIC_CLOUDFLARE_USE_TURNSTILE_IN_BOOKER === "1") {
+    await checkCfTurnstileToken({
+      token: req.body["cfToken"] as string,
+      remoteIp: userIp,
+    });
+  }
+
+  if (process.env.NEXT_PUBLIC_VERCEL_USE_BOTID_IN_BOOKER === "1") {
+    const verification = await checkBotId({
+      developmentOptions: {
+        bypass: "BAD-BOT", // default: 'HUMAN'
+      },
+    });
+
+    if (verification.isBot) {
+      return NextResponse.json({ error: "Access denied" }, { status: 403 });
+    }
+  }
+
+  await checkRateLimitAndThrowError({
+    rateLimitingType: "core",
+    identifier: piiHasher.hash(userIp),
+  });
+
+  const session = await getServerSession({ req });
+  /* To mimic API behavior and comply with types */
+  req.body = {
+    ...req.body,
+    creationSource: CreationSource.WEBAPP,
+  };
+
+  const booking = await handleNewBooking({
+    bookingData: req.body,
+    userId: session?.user?.id || -1,
+    hostname: req.headers.host || "",
+    forcedSlug: req.headers["x-cal-force-slug"] as string | undefined,
+  });
+  // const booking = await createBookingThroughFactory();
+  return booking;
+
+  //  To be added in the follow-up PR
+  // async function createBookingThroughFactory() {
+  //   console.log("Creating booking through factory");
+  //   const regularBookingService = getRegularBookingService();
+
+  //   const booking = await regularBookingService.createBooking({
+  //     bookingData: req.body,
+  //     bookingMeta: {
+  //       userId: session?.user?.id || -1,
+  //       hostname: req.headers.host || "",
+  //       forcedSlug: req.headers["x-cal-force-slug"] as string | undefined,
+  //     },
+  //   });
+  //   return booking;
+  // }
+}
+
+export default defaultResponder(handler, "/api/book/event");
